Fix yandex
