+++ core/src/services/yandex_disk/mod.rs
@@ -22,5 +22,4 @@ pub use backend::YandexDiskConfig;
 mod core;
 mod error;
 mod lister;
-mod reader;
 mod writer;
